api_yamdb/requirements.txt
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
colorama==0.4.4
Django==2.2.16
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-csv==2.1.1
djangorestframework-simplejwt==5.0.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
py==1.10.0
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
unicodecsv==0.14.1
urllib3==1.26.7
